add sidebar

+++ src/actions/auth.actions.js
@@ -50,18 +50,26 @@ export const isUserLoggedIn = () => {
       dispatch({
         type: authConstants.LOGIN_FAILURE,
         payload: {
-          error: 'Failed to login',
-        }
+          error: "Failed to login",
+        },
       });
     }
   };
 };
 
-export const signout =()=>{
-  return async dispatch =>{
-    localStorage.clear();
-    dispatch({
-      type:authConstants.LOGOUT_REQUEST
-    });
-  }
-}
+export const signout = () => {
+  return async (dispatch) => {
+    dispatch({ type: authConstants.LOGOUT_REQUEST });
+    const res = await axios.post(`/admin/signout`);
+
+    if (res.status === 200) {  
+      localStorage.clear();
+      dispatch({ type: authConstants.LOGOUT_SUCCESS });
+    } else {
+      dispatch({
+        type: authConstants.LOGOUT_FAILURE,
+        payload: { error: res.data.error },
+      });
+    }
+  };
+};
